Fix reference counting
I don't know how this could possibly work, but if you create a new application, add this lib and just try to BINDO with -observe: it won't work, because right after the creation of the binding it gets deallocated: `BNDBinding` is not retained. You had an `NSMutableSet`, where you add `BNDBinding`s, but now it's gone and only `BNDBindingKVOObserver` are retained by association. But they have a weak reference to parent `BNDBinding`. That's a fail.

To fix this we must make `BNDBindingKVOObserver`s to hold strong reference to parent `BNDBinding`.